Move to Castle.MicroKernel.Internal
#565
