+++ src/Castle.Facilities.WcfIntegration/Client/Discovery/DiscoveryProxyCache.cs
@@ -16,7 +16,8 @@ namespace Castle.Facilities.WcfIntegration
 {
 	using System;
 	using System.ServiceModel.Discovery;
-	using Castle.Core.Internal.Locking;
+
+	using Castle.MicroKernel.Internal;
 
 	public class DiscoveryProxyCache : DiscoveryEndpointProvider, IDisposable
 	{
