+++ src/Castle.Facilities.WcfIntegration/Service/Discovery/LoadBalance/ListBasedLoadBalancePolicy.cs
@@ -17,7 +17,8 @@ namespace Castle.Facilities.WcfIntegration
 	using System;
 	using System.Collections.Generic;
 	using System.ServiceModel.Discovery;
-	using Castle.Core.Internal.Locking;
+
+	using Castle.MicroKernel.Internal;
 
 	public delegate bool PolicyMembership(EndpointDiscoveryMetadata endpoint);
