+++ src/Castle.Windsor/MicroKernel/Internal/ILockHolder.cs
@@ -12,15 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Castle.Core.Internal.Locking
+namespace Castle.MicroKernel.Internal
 {
 	using System;
-	using System.ComponentModel;
 
-	[EditorBrowsable(EditorBrowsableState.Never)]
-	public interface IUpgradeableLockHolder : ILockHolder
+	public interface ILockHolder : IDisposable
 	{
-		ILockHolder Upgrade();
-		ILockHolder Upgrade(bool waitForLock);
+		bool LockAcquired { get; }
 	}
 }
