+++ src/Castle.Windsor/MicroKernel/Internal/IUpgradeableLockHolder.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2004-2010 Castle Project - http://www.castleproject.org/
+﻿// Copyright 2004-2009 Castle Project - http://www.castleproject.org/
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,20 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Castle.Core.Internal.Locking
+namespace Castle.MicroKernel.Internal
 {
 	using System.ComponentModel;
 
 	[EditorBrowsable(EditorBrowsableState.Never)]
-	internal class NoOpLock : ILockHolder
+	public interface IUpgradeableLockHolder : ILockHolder
 	{
-		public static readonly ILockHolder Lock = new NoOpLock();
-
-		public void Dispose()
-		{
-
-		}
-
-		public bool LockAcquired => true;
+		ILockHolder Upgrade();
+		ILockHolder Upgrade(bool waitForLock);
 	}
 }
