+++ src/Castle.Windsor/MicroKernel/Internal/SlimReadWriteLock.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Castle.Core.Internal.Locking
+namespace Castle.MicroKernel.Internal
 {
 	using System.Threading;
