+++ src/Castle.Windsor/MicroKernel/Internal/SlimUpgradeableReadLockHolder.cs
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Castle.Core.Internal.Locking
+namespace Castle.MicroKernel.Internal
 {
 	using System.Threading;
-	
+
 	internal class SlimUpgradeableReadLockHolder : IUpgradeableLockHolder
 	{
 		private readonly ReaderWriterLockSlim locker;
